Default to one job per fork if the environment variable isn't specified.
I feel it's better to default to existing functionality (ie a single job
per fork) rather than crashing. Once we're at the point where the
exception is getting raised we've already pulled the job from the queue
and by crashing it will get released into the ether.

I feel it's better to error on the side of not losing work.