Exit the child process if we run out of work to do.
Previously, if we had JOBS_PER_FORK set to 20 and 1 job in the queue
Resque would fork a child to handle that one job and then the plugin
would hit redis 19 more times as quickly as possible to get the next
job.

This change causes the worker to return from the fork if we run out of
work to instead of potentially hammering redis.